Clarify grammar for UTC offset strings
This commit adds productions and prose to more clearly define the
syntax and use of UTC Offset strings in ECMAScript.

`|NormalizedUTCOffset|` defines the unambiguous, canonical form
of a UTC offset string. It's used to provide a deterministic format for
offset strings generated by ECMAScript. It wil be used in a subsequent
normative commit in this PR. It will also used in Temporal.

`|DateUTCOffset|` is the subset of `|NormalizedUTCOffset|` that is
accepted by `Date.parse`. It replaces existing prose in
https://tc39.es/ecma262/#sec-date-time-string-format.