Normative: assert SystemTimeZoneIdentifier offsets
SystemTimeZoneIdentifier currently supports both named and offset
time zones. Named time zone identifiers are always asserted to be in
canonical form, but when offset string support was added, we forgot
to include a corresponding step to ensure that offset strings are
normalized too.

This commit adds an assertion to fix this oversight. If the system time
zone is an offset time zone, then SystemTimeZoneIdentifier now asserts
that the offset string is normalized into its canonical form.

AFAICT there are no major implementations that currently return
offset time zones from this AO. (I tested many implementations including
Chrome/Firefox/Node on MacOS/Linux/Windows, as well as Safari on
MacOS.) So now is a good time to fix this AO because we won't break
existing ECMAScript programs.